Major Update Made a new Joystick controller in OI mapped buttons to the MoveToLevel command Changed arm control to 1 command Fixed decrement and increment cap Fixed offset in levels made 2 commands for piston on gameMech to use on controllers Created command to spin flywheels for set speed and duration Overall cleanup. (Gabe still hates my coding style) All the best, MJ :/

+++ src/main/java/frc/robot/subsystems/Arm.java
@@ -30,7 +30,7 @@ public class Arm extends Subsystem {
         setDefaultCommand(new Levels());
     }
 
-    public void rotateToPosition(int desiredtargetEncoderValue) {
+    public void rotateToPosition(double desiredtargetEncoderValue) {
         // checks if the target has changed
         // if it has changed, reset the base variables to 0;
         if (desiredtargetEncoderValue != previousDesiredtargetEncoderValue) {
